+++ lib/tailwind_ui.dart
@@ -1,5 +1,6 @@
 library tailwind_ui;
 
+export 'src/application_ui/elements/avatars/avatar_group_bottom_to_top.dart';
 export 'src/application_ui/elements/avatars/circular_avatar.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_notification.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_placeholder_icon.dart';