+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rounded avatar with notification component
 - Circular avatar with placeholder icon component
 - Circular avatar with placeholder initials component
+- Avatar group bottom to top component